refactor(rpc-server): Query method logging improvements
